% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{as_age_groups}
\alias{as_age_groups}
\title{Categorise continuous age into age groups}
\usage{
as_age_groups(
  var,
  min = 0,
  max = 100,
  by = 10,
  grouping_method = c("user_defined", "ons_1", "ons_2", "nhs_survey")
)
}
\arguments{
\item{var}{Variable name}

\item{min}{Numeric, minimum age}

\item{max}{Numeric, maximum age}

\item{by}{Numeric, increment of the age categories}

\item{method}{}
}
\value{

}
\description{
Categorise continuous age into age groups
}
