% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{calc_monthly_freq}
\alias{calc_monthly_freq}
\title{Calculate frequencies of groups per month}
\usage{
calc_monthly_freq(
  data,
  by,
  date_var_name,
  .drop_latest_month = TRUE,
  .calc_year_month_day_vars = TRUE
)
}
\arguments{
\item{data}{Data}

\item{by}{Grouping variables used in group_by}

\item{date_var_name}{String specifying date variable, needs to be class date}

\item{.drop_latest_month}{Logical, specifying whether or not to drop the most recent month}

\item{.calc_year_month_day_vars}{Logical, specifying whether to calculate separate year, month, and day variable}
}
\description{
Calculate frequencies of groups per month
}
