% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{ds_theme}
\alias{ds_theme}
\title{NOTTSHC CDU data science team ggplot2 theme details}
\usage{
ds_theme(left_label = TRUE)
}
\arguments{
\item{left_label}{Logical. Indicates whether to include the x axis labels and
ticks which appear on the left of the chart. Default is to include and useful to exclude
on charts specifically needed to be 'clean' for example Smoking Prevalence.}
}
\value{

}
\description{
Inspired from the chart layout
used by TraffordDataLab https://github.com/traffordDataLab/assets/tree/master/theme/ggplot2
}
\details{
A ggplot2 theme with extended details
}
\section{Last updated by}{

Zoë Turner
}

\section{Last updated date}{

2021-04-01
}

\examples{
# load some data
df <- data.frame(
 religion = c("Christian", "Buddhist", "Hindu", "Jewish", "Muslim", "Sikh",
 "Other Religion", "No Religion", "Not Stated"),
 count = c(143639, 768, 2271, 2413, 12994, 1652, 566, 47968, 14307)
)

# create a ggplot object
plot <- df \%>\%
 dplyr::arrange(count) \%>\%
 dplyr::mutate(religion = factor(religion,
                                 levels = religion)) \%>\%
 ggplot2::ggplot(ggplot2::aes(religion, count)) +
 ggplot2::geom_col(fill = "#fc6721",
                   alpha = 0.8,
                   show.legend = FALSE) +
 ggplot2::coord_flip() +
 ggplot2::labs(x = NULL, y = "Residents",
               title = "This is the title of the chart",
               subtitle = "This is the subtitle",
               caption = "Source: Table KS209EW, Census 2011  |  @DataScienceNott")

# style with the ds_theme putting y axis label and ticks back
plot + nottshcMethods::ds_theme(left_label = FALSE)
}
