% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{ggplotly_nottshc}
\alias{ggplotly_nottshc}
\title{Convert ggplot object into plotly}
\usage{
ggplotly_nottshc(
  ggplot,
  xtitle = NULL,
  ytitle = NULL,
  aes_txt_tooltip = TRUE,
  display_mode_bar = FALSE
)
}
\arguments{
\item{ggplot}{ggplot2 object}

\item{xtitle}{String, x lab title}

\item{ytitle}{String, y lab title}

\item{aes_txt_tooltip}{Logical, specifying whether or not to use tooltip}

\item{display_mode_bar}{Logical, specifying whether to show the annoying
mode bar from plotly, FALSE by default!}
}
\value{

}
\description{
There are some problems with x and y axis labels when converting
ggplot2 to plotly, especially when using facets. This fun has a workaround that is not perfect.
}
