% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash.R
\name{hash}
\alias{hash}
\title{One way hash a vector}
\usage{
hash(x, type = "md5", n_char = NULL)
}
\arguments{
\item{x}{vector of numbers or characters}

\item{type}{hashing algorithm to use}

\item{n_char}{number of characters to return}
}
\value{
one way hashed vector of the same length
}
\description{
One way hash a vector
}
\examples{
hash("This is secret stuff")
hash(c("Also handles", NA, "NA values"))
}
