% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_vector.R
\name{sample_vector}
\alias{sample_vector}
\title{CDU data science team ggplot2 theme}
\usage{
sample_vector(values, weights, length)
}
\arguments{
\item{values}{Vector- the thing you want to sample from}

\item{weights}{vector of integers the same length as values- the proportions
with which you want to sample}

\item{length}{integer- how long you want the return value to be- e.g. the
number of rows in a dataframe}
}
\description{
A ggplot2 theme with a white panel background, no grid lines,
large axis and legend titles,
and increased text padding for better readability.
}
\examples{
sample_vector(values = c(NA, "Male", "Female", "Other"),
weights = c(10, 50, 50, 2),
length = 100)

}
