% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_helper.R
\name{year_date}
\alias{year_date}
\title{Function to give the 'yyyy' string to a date column (of formats date, int and character)}
\usage{
year_date(data, date_col, type = c("earliest", "latest"))
}
\arguments{
\item{data}{Dataset}

\item{date_col}{Date, the date column to find the latest/earliest. If this
an integer (sk synthetic key) converts to date format}

\item{type}{String, latest gives the most recent year in the column whilst earliest gives the
first}
}
\value{

}
\description{
Function to give the 'yyyy' string to a date column (of formats date, int and character)
}
\examples{
test <- data.frame(
 date = sample(seq(as.Date('2007/01/01'), as.Date('2021/01/01'), by = "day"), 12)) \%>\%
 dplyr::mutate(date_sk = as.integer(format(as.Date(date), "\%Y\%m\%d"))) \%>\%
 dplyr::mutate(charc_date = as.character(date))

year_date(test, date, type = "latest")
year_date(test, date, type = "earliest")

year_date(test, date_sk, type = "latest")
year_date(test, date_sk, type = "earliest")

year_date(test, charc_date, type = "latest")
year_date(test, charc_date, type = "earliest")

}
