% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{as_age_groups}
\alias{as_age_groups}
\title{Categorise continuous age into age groups}
\usage{
as_age_groups(
  var,
  min = 0,
  max = 100,
  by = 10,
  grouping_method = c("user_defined", "ons_1", "ons_2", "nhs_survey")
)
}
\arguments{
\item{var}{Name of variable or vector}

\item{min}{Numeric, specifying the minimum age of the first age group.
Age values lower than this will be returned as missing values (NA)}

\item{max}{Numeric, specifying the upper end of the last age group}

\item{by}{Numeric, increment of the age categories}

\item{method}{String, specyfing the method to be used for grouping age into categories.
Details about the different methods are available here:
"user_defined", "ons_1", "ons_2", "nhs_survey" ... TODO}
}
\value{

}
\description{
Categorise continuous age into age groups
}
\examples{
# Example using a vector:
set.seed(123)
age <- sample(1:100, 100, replace = T)
as_age_groups(age)

# Example using a data frame
tibble::tibble(age = sample(1:115, 100, replace = T)) \%>\%
  dplyr::mutate(age_groups = as_age_groups(age, min = 10, max = 100))
}
