% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{as_age_groups}
\alias{as_age_groups}
\title{Categorise continuous age into age groups}
\usage{
as_age_groups(var, min = 0, max = 100, by = 10)
}
\arguments{
\item{var}{Variable name}

\item{min}{Numeric, minimum age}

\item{max}{Numeric, maximum age}

\item{by}{Numeric, increment of the age categories}
}
\value{

}
\description{
Categorise continuous age into age groups
}
